Add geoLocation, geoLocationPoint, and geoLocationBox models and MySQL tables
GeoLocation has a polymorphic association with geoLocationPoint and
geoLocationBox. A geoLocationBox or geoLocationPoint can have many
geoLocations, and a record can have one geoLocation. The MySQL tables
for Points and Boxes make use of the rgeo, rgeo-activerecord, and
activerecord-mysql2spatial-adapter gems. This allows the table to
store coordinates as the spatial data types 'point' and 'polygon'.
Rgeo may require the use of the GEOS library
(http://trac.osgeo.org/geos).