Fix enforcement of map drawing limitations
The previous system of "grey-ing out" the drawing tool was buggy.
Instead, JQuery toggles the visibility of the Draw toolbar itself
when the limit of geospatial objects on the map has been reached.

Also, the geoLocationBox form fields are now updated when the
drawn rectangle is edited. Entering coordinates into the form
will trigger the creation of a new rectangle if one did not
previously exist.