Set Point fields on marker creation
When a marker is created, it is connected to the first Point's
fields via hash variable. The Point's lat and lng fields are then
set to the marker's coordinates. Currently, each marker created
will overwrite the previous marker's coordinates in the first
Point's fields.

Markers are highlighted using a differently colored icon image
when moused over.