Add new coordinate fields when a marker is created
Rather than try to link markers with existing Point fields, the
submission form is now set to create new Point fields on
successful marker creation. These new fields are the same as the
regular Points fields, except these cannot be edited by the user,
and are tied to the original coordinates of the marker.

(However, editing the marker will not yet affect the associated
Point fields.)