Completed hybrid Points creation
When Point fields or markers are indicated for removal, the draw
tool will be shown or hidden accordingly. Also, the button to add
new Point fields is disabled whenever the limit on the number of
Points has been reached. When a marker is deleted, its corresponding
disabled Point fields will also be removed.