Fix number of fields created alongside a Marker
Every time the checkbox for geographic metadata was toggled,
creating a Marker would create +1 extra Point coordinate pair.
Now, whenever the checkbox is clicked off, the previous Point
fields are destroyed, the markerMap is cleared, the draw/edit
toolbars are removed, and the map layers are cleared of objects.

Also, validation for geoLocationBox has been fixed to reflect
the fact that "validates_associated" does not work with "has_one"
associations. This solves a bug which always saved a Box, even
with NULL fields.